+++ labstep/entities/resource/model.py
@@ -6,7 +6,7 @@ from labstep.service.helpers import getTime
 from labstep.generic.entityPrimary.model import EntityPrimary
 from labstep.generic.entityWithMetadata.model import EntityWithMetadata
 from labstep.entities.resourceItem.model import ResourceItem
-from labstep.constants import UNSPECIFIED
+from labstep.constants import PLACEHOLDER_SVG, UNSPECIFIED
 
 
 class Resource(EntityPrimary, EntityWithMetadata):
@@ -325,3 +325,136 @@ class Resource(EntityPrimary, EntityWithMetadata):
         import labstep.entities.experimentDataField.repository as experimentDataFieldRepository
 
         return experimentDataFieldRepository.getDataFields(self)
+
+    def addChemicalMetadata(self, structure=UNSPECIFIED, iupac_name=UNSPECIFIED, cas=UNSPECIFIED, density=UNSPECIFIED, inchi=UNSPECIFIED, molecular_formula=UNSPECIFIED, molecular_weight=UNSPECIFIED, smiles=UNSPECIFIED, safety={}, svg=UNSPECIFIED):
+        """
+        Add chemical metadata to a Resource.
+
+        Parameters
+        ----------
+        structure (str)
+            The chemical structure as SMILES or MOL
+        iupac_name (str)
+            The IUPAC Name
+        cas (str)
+            The CAS Number
+        denstity (str)
+            The density of the chemicals
+        inchi (str)
+            The InChI of the chemical
+        molecular_formula (str)
+            The Molecular Formula
+        molecular_weight (str)
+            The Molecular Weight
+        smiles (str)
+            The SMILES string describing the chemical
+        svg (str)
+            An SVG preview of the chemical structure
+
+        Returns
+        -------
+        :class:`~labstep.entities.metadata.model.Metadata`]
+
+        Example
+        -------
+        ::
+
+            my_resource = user.getResource(17000)
+            my_resource.addChemicalMetadata(
+                            structure='C1=CC=CC=C1',
+                            iupac_name="benzene",
+                            cas="71-43-2",
+                            density="0.879 at 68 °F (USCG, 1999)",
+                            inchi="InChI=1S/C6H6/c1-2-4-6-5-3-1/h1-6H",
+                            molecular_formula="C6H6",
+                            molecular_weight="78.11",
+                            smiles='C1=CC=CC=C1',
+                            safety={
+                                 "precautionary": {
+                                     "General": [],
+                                     "Prevention": [
+                                         "P203",
+                                     ],
+                                     "Storage": [
+                                         "P403+P235",
+                                     ],
+                                     "Disposal": [
+                                         "P501"
+                                     ]
+                                 },
+                                 "ghs_hazard_statements": [
+                                     "H372"
+                                 ],
+                                 "pictograms": [
+                                     "GHS02",
+                                 ]
+                             }
+                            )
+        """
+        metadata = {
+            "IUPACName": iupac_name if iupac_name is not UNSPECIFIED else None,
+            "InChI": inchi if inchi is not UNSPECIFIED else None,
+            "MolecularFormula": molecular_formula if molecular_formula is not UNSPECIFIED else None,
+            "CAS": cas if cas is not UNSPECIFIED else None,
+            "Density": density if density is not UNSPECIFIED else None,
+            "SMILES": smiles if smiles is not UNSPECIFIED else None,
+            "MolecularWeight": molecular_weight if molecular_weight is not UNSPECIFIED else None,
+            "Safety": safety if safety is not UNSPECIFIED else {}
+        }
+
+        return self.addMetadata('Structure', 'molecule', extraParams={'molecule': {
+            'name': 'Untitled',
+            'data': structure if structure is not UNSPECIFIED else smiles,
+            'svg': svg if svg is not UNSPECIFIED else PLACEHOLDER_SVG,
+            'pubchem': {
+                'reactants': [metadata],
+                'products': []
+            }
+        }})
+
+    def getChemicalMetadata(self):
+        '''
+        Returns the chemical metadata associated with the resource
+
+        Returns
+        -------
+        {
+            "Structure": string,
+            "IUPACName": string,
+            "InChI": string ,
+            "MolecularFormula": string,
+            "CAS": string,
+            "Density": string,
+            "SMILES": string,
+            "MolecularWeight": string,
+            "Safety": {
+                "precautionary": {
+                    "General": [string],
+                    "Prevention": [string],
+                    "Response": [string],
+                    "Storage": [string],
+                    "Disposal": [string]
+                },
+                "ghs_hazard_statements": [string],
+                "pictograms": [string]
+            }
+            "SVG": string
+        }
+
+        Example
+        -------
+        ::
+
+            my_resource = user.getResource(17000)
+            my_resource.getChemicalMetadata()
+        '''
+
+        self.update()
+
+        metadata = self.molecule['pubchem']['reactants'][0]
+
+        metadata['Structure'] = self.molecule['data']
+
+        metadata['SVG'] = self.molecule['svg']
+
+        return metadata