+++ CHANGELOG.md
@@ -4,6 +4,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (tries to) adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# 3.16.0 2022-04-29
+## Added
+- `ChemicalReaction` model
+- `Chemical` model
+- `addChemicalReaction` and `getChemicalReactions` methods on `Experiment` and `ExperimentProtocol`
+- `addChemicalMetadata` and `getChemicalMetadata` methods on `Resource`
+
+## Fixed
+- `getFile` method can now get deleted files
+# 3.15.2 2022-04-19
+## Fixed
+- Export bugs.
+  
+# 3.15.1 2022-04-19
+## Fixed
+- Export foldering name bug.
+# 3.15.0 2022-04-14
+## Added
+- `getUserAgent` and `setUserAgent` in `labstep.service.config.configService`
+# 3.14.0 2022-04-04
+
+## Added
+- New Entity `Molecule`
+## Changed
+- Entity name for `ExperimentInventoryField`
+## Fixed
+- Setting value for options field bug.
+
+# 3.13.0 2022-02-18
+
+## Added
+- Include PDF in experiment export
+
 ## 3.12.0 2022-02-18
 
 - `ExperimentMaterial` class renamed to `ExperimentInventoryField`